convert button stories to tsx
